requirements.txt
# 视频吸引力预测模型 - 依赖包列表
# 
# 安装方式：
#   pip install -r requirements.txt
#
# 如果使用conda：
#   conda install pytorch torchvision -c pytorch
#   pip install -r requirements.txt

# PyTorch (根据你的系统选择合适的版本)
# CPU版本:
# pip install torch torchvision
# 
# CUDA版本 (如果有NVIDIA GPU):
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# Apple Silicon MPS版本:
# pip install torch torchvision

torch>=2.0.0
torchvision>=0.15.0

# Transformers (用于BERT模型)
transformers>=4.30.0

# 数据处理
pandas>=1.5.0
numpy>=1.23.0
Pillow>=9.0.0

# 机器学习工具
scikit-learn>=1.2.0

# 可视化
matplotlib>=3.6.0

# 进度条（可选）
tqdm>=4.64.0
